==> Modeling/9_CF_ALS_Model.ipynb <==
This code uses the training, validation and test sets created on the notebook: Implicit_Rating_Calculation_category.ipynb to create an Collaborative Filtering (CF) model using Alternating Least Squares (ALS) from Pyspark.

# Loading needed pyspark libraries
from pyspark.ml.recommendation import ALS
from pyspark.ml.evaluation import RegressionEvaluator
from pyspark.sql import functions as F
from pyspark.sql.types import IntegerType
from pyspark.ml.tuning import CrossValidator, ParamGridBuilder, TrainValidationSplit
from pyspark.ml.feature import StringIndexer

# Initializing spark context and setting up checkpoint directory
sc = spark.sparkContext
sc.setCheckpointDir("s3a://myaws-capstone-bucket/data/modeling/checkpointdir")

# Reading train, validation and test datasets for implicit_cat data
train_file ='s3://myaws-capstone-bucket/data/modeling/input/implicit_cat_rating_train.csv'
df_train =spark.read.option("header","true").csv(train_file)

val_train_file ='s3://myaws-capstone-bucket/data/modeling/input/implicit_cat_rating_val_train.csv'
df_train_val =spark.read.option("header","true").csv(val_train_file)

test_file ='s3://myaws-capstone-bucket/data/modeling/input/implicit_cat_rating_test.csv'
df_test =spark.read.option("header","true").csv(test_file)

val_test_file ='s3://myaws-capstone-bucket/data/modeling/input/implicit_cat_rating_val_test.csv'
df_test_val =spark.read.option("header","true").csv(val_test_file)

df_train.show()

# Splitting product column to obtain numeric id since PySpark expects all columns for model to be integer type

# Transforming user_id and catID and rating to integer types since PySpark expects all columns for model to be integer type
df_train = df_train.withColumn("user_id", df_train["user_id"].cast(IntegerType()))
df_train = df_train.withColumn("catID", df_train["catID"].cast(IntegerType()))
df_train = df_train.withColumn("implicit_rating", df_train["implicit_rating"].cast(IntegerType()))

df_train.show()

# Checking type of columns
df_train.dtypes

# Repeating process for validation and testing dataframes


# Transforming user_id and catID and rating to integer types since PySpark expects all columns for model to be integer type
df_train_val = df_train_val.withColumn("user_id", df_train_val["user_id"].cast(IntegerType()))
df_train_val = df_train_val.withColumn("catID", df_train_val["catID"].cast(IntegerType()))
df_train_val = df_train_val.withColumn("implicit_rating", df_train_val["implicit_rating"].cast(IntegerType()))

# Transforming user_id and catID and rating to integer types since PySpark expects all columns for model to be integer type
df_test = df_test.withColumn("user_id", df_test["user_id"].cast(IntegerType()))
df_test = df_test.withColumn("catID", df_test["catID"].cast(IntegerType()))
df_test = df_test.withColumn("implicit_rating", df_test["implicit_rating"].cast(IntegerType()))

# Transforming user_id and catID, and rating to integer types since PySpark expects all columns for model to be integer type
df_test_val = df_test_val.withColumn("user_id", df_test_val["user_id"].cast(IntegerType()))
df_test_val = df_test_val.withColumn("catID", df_test_val["catID"].cast(IntegerType()))
df_test_val = df_test_val.withColumn("implicit_rating", df_test_val["implicit_rating"].cast(IntegerType()))

df_train_val.show()

# Building a generic ALS CF model 
als = ALS(userCol="user_id", itemCol="catID", ratingCol="implicit_rating",
          coldStartStrategy="drop", nonnegative = True, implicitPrefs = False)

# Performing hyperparameter tuning with ParamGrid that will try for each of the different parameter combinations
param_grid = ParamGridBuilder().addGrid(als.rank, [1, 5, 10]).addGrid(als.maxIter, [5, 10, 20]).addGrid(als.regParam, [0.05, 0.1, 0.5]).build()

# Evaluator to evaluate the performance of the model
evaluator = RegressionEvaluator(metricName="rmse", labelCol="implicit_rating", predictionCol="prediction")

# Using CrossValidator to perform cross validation 
cv = CrossValidator(estimator = als,
estimatorParamMaps = param_grid,
evaluator = evaluator,
numFolds = 5)# Performing 5 folds

# Running the CV on the validation training dataset
model = cv.fit(df_train_val)

# Extracting the best model from cross validation
best_model = model.bestModel

# Make predictions on the validation test set and evaluate using RMSE
predictions = best_model.transform(df_test_val)
rmse = evaluator.evaluate(predictions)

# Obtaining the evaluation metrics and model parameters
print ("Best Model Parameters")
print ("RMSE: " + str(rmse))
print (" Rank: " + str(best_model.rank))
print (" Max Iter: " + str(best_model._java_obj.parent().getMaxIter()))
print (" RegParam: " + str(best_model._java_obj.parent().getRegParam())) 

# Building the recommendation model using ALS best parameter model
als = ALS(maxIter=20, rank= 1, regParam=0.1, implicitPrefs=False,coldStartStrategy="drop",
          userCol='user_id', itemCol='catID', ratingCol='implicit_rating')

model = als.fit(df_train)
predictions = model.transform(df_test)
rmse = evaluator.evaluate(predictions)
print("Root-mean-square error = " + str(rmse))

# Generating top 10 recommendations for each user
recommendations = model.recommendForAllUsers(10)

# Viewing recommendations df schema to know which columns to select when formatting the df
recommendations.printSchema()

# Formatting results
recommendations_output = recommendations.select('user_id',F.posexplode('recommendations')).select('user_id','col.catID','pos','col.rating').checkpoint()
recommendations_output.show()

# Saving results to S3 - Using coalesce function to merge all the small files Spark creates into one so that there is only one final file
recommendations_output.coalesce(1).write.option("quote","\u0000").csv("s3a://myaws-capstone-bucket/data/modeling/ALS/results",mode="overwrite",sep=',')
