==> Modeling/11_CNN_Spotlight_param2.ipynb <==
This notebook uses the package spotlight found at: https://github.com/maciejkula/spotlight, to create a CNN recommendation model. Information on the model can be found in the below articles: 
- https://towardsdatascience.com/introduction-to-recommender-system-part-2-adoption-of-neural-network-831972c4cbf7
- https://towardsdatascience.com/how-to-build-powerful-deep-recommender-systems-using-spotlight-ec11198c173c 

Also using as reference the documentation github page: https://github.com/maciejkula/spotlight/tree/master/examples/movielens_sequence

This code uses the training, validation and test sets created on the notebook: Implicit_Rating_Calculation_category.ipynb to train and test the model.

#### Installing the Spotlight library 

!pip install git+https://github.com/maciejkula/spotlight.git@master --upgrade 

# Loading needed libraries
import numpy as np
import pandas as pd
import datetime as dt
from datetime import date
import torch
from sklearn.preprocessing import LabelEncoder
import gc

# Spotlight Libraries
from spotlight.sequence.implicit import ImplicitSequenceModel
from spotlight.sequence.representations import CNNNet
from spotlight.interactions import Interactions
from spotlight.cross_validation import random_train_test_split
from spotlight.evaluation import sequence_mrr_score

# Loading libraries for S3 bucket connection
import boto3
import io
from io import StringIO,BytesIO, TextIOWrapper
import gzip

client = boto3.client('s3') 
resource = boto3.resource('s3') 

The expected data input should contain the user ids, the product ids, implicit rating and the timestamps (which are optional)

#### Data Preparation

# Reading Training,validation and testing dfs - Using the created T_implicit_cat data with timestamp
train_df = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/input/T_implicit_cat_rating_train.csv')
test_df = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/input/T_implicit_cat_rating_test.csv')

# Transforming event_time columns into timestamp  
train_df['event_date'] = train_df['event_time'].str[:19]# Grabbing only timestamp portion from original event_time column
train_df['event_date'] = train_df['event_date'].astype('datetime64[ns]')
train_df['timestamp'] = train_df['event_date'].values.astype(np.int64)//10**9
train_df['timestamp'] = train_df['timestamp'].astype(np.int32)
train_df.head()

# Transform category and user ids to needed format

# instantiating the labelencoder object
le = LabelEncoder()

train_df['catID'] = train_df['catID'].astype(np.int32)+1
train_df['userID'] = le.fit_transform(train_df['user_id'])
train_df['userID'] = train_df['userID'].astype(np.int32)+1

# Applying the same to test df

test_df['event_date'] = test_df['event_time'].str[:19]# Grabbing only timestamp portion from original event_time column
test_df['event_date'] = test_df['event_date'].astype('datetime64[ns]')
test_df['timestamp'] = test_df['event_date'].values.astype(np.int64)//10**9
test_df['timestamp'] = test_df['timestamp'].astype('str')

# Transform product and user ids to needed format
test_df['catID'] = test_df['catID'].astype(np.int32)+1
test_df['userID'] = le.fit_transform(test_df['user_id'])
test_df['userID'] = test_df['userID'].astype(np.int32)+1

test_df.head()

# Creating interaction Spotlight objects since Spotlight model expects this specific type of object
train=Interactions(user_ids=train_df['userID'].to_numpy(),item_ids=train_df['catID'].to_numpy(),timestamps=train_df['timestamp'].to_numpy())

test=Interactions(user_ids=test_df['userID'].to_numpy(),item_ids=test_df['catID'].to_numpy(),timestamps=test_df['timestamp'].to_numpy())

# Setting min and max sequence number
# Based on analysis done on notebook:Implicit_Rating_Calculation_final.ipynb we have set a minimum of 2 and max of 8 sequences
min_sequence_length = 2
max_sequence_length = 8
random_state = np.random.RandomState(572)

train_seq = train.to_sequence(max_sequence_length=max_sequence_length,
                              min_sequence_length=min_sequence_length)

test_seq = test.to_sequence(max_sequence_length=max_sequence_length,
                              min_sequence_length=min_sequence_length)

hyperparameters = {
        'n_iter': 5,
        'batch_size': 64,
        'l2': 0.0,
        'learning_rate': 1e-3,
        'loss': 'adaptive_hinge',
        'embedding_dim': 32,
        'kernel_width':5,
        'dilation':[1, 2, 4, 8, 16],
        'num_layers':5,
        'nonlinearity':"relu",
        'residual':True
    }


net = CNNNet(train.num_items,
             embedding_dim=hyperparameters['embedding_dim'],
             kernel_width=hyperparameters['kernel_width'],
             dilation=hyperparameters['dilation'],
             num_layers=hyperparameters['num_layers'],
             nonlinearity=hyperparameters['nonlinearity'],
             residual_connections=hyperparameters['residual'])

model = ImplicitSequenceModel(loss=hyperparameters['loss'],
                              representation=net,
                              batch_size=hyperparameters['batch_size'],
                              learning_rate=hyperparameters['learning_rate'],
                              l2=hyperparameters['l2'],
                              n_iter=hyperparameters['n_iter'],
                              use_cuda=torch.cuda.is_available(),
                              random_state=random_state)

model.fit(train_seq, verbose=True)

torch.save(model, 'CNN.pt')

sequences = test_seq.sequences

pred_list=[]
for i in range(len(sequences)):
    predictions = model.predict(sequences[i])
    pred_list.append(predictions.argsort()[-10:][::-1])

# Converting prediction of CNN into a df
predictions_df = pd.DataFrame(data=pred_list)
predictions_df.head()

# Creating a df of user ids to have them in order
users_df = pd.DataFrame(data=test_seq.user_ids)
users_df.columns = ['userID']

predictions_df['userID']= users_df['userID']

predictions_df.head()

predictions_df = predictions_df.groupby('userID').tail(1)# Handling cases with multiple sequences to only grab latest one

# Rearranging recs from rows to columns
predictions_df = predictions_df.melt(id_vars=["userID"], var_name="Category_Rank", value_name="catID")
predictions_df.head()

cat_mapping = test_df[['catID','category','category_id']]
cat_mapping = cat_mapping.drop_duplicates(subset=['catID','category','category_id'])

user_mapping = test_df[['userID','user_id']]
user_mapping = user_mapping.drop_duplicates(subset=['userID','user_id'])

# Merging predictions_df to obtain the correct user and category
predictions_df = pd.merge(predictions_df, cat_mapping,  how='inner', on='catID')
predictions_df = pd.merge(predictions_df, user_mapping,  how='inner', on='userID')
#Dropping duplicates
predictions_df = predictions_df.drop_duplicates(['user_id','catID','category','category_id'])
predictions_df.head(10)

# Saving Results in S3
predictions_df.to_csv('s3://myaws-capstone-bucket/data/modeling/output/CNN_param2.csv',index=False)

predictions_df.nunique()
