==> Modeling/17_MF_Spotlight_Param2.ipynb <==
This notebook uses the package spotlight found at: https://github.com/maciejkula/spotlight, to create an Matrix factorization (MF) recommendation system using BilinearNet. Information on the model can be found here: https://github.com/maciejkula/spotlight/tree/master/examples/movielens_explicit 

Also using as reference the following article: https://towardsdatascience.com/how-to-build-powerful-deep-recommender-systems-using-spotlight-ec11198c173c

This code uses the training, validation and test sets created on the notebook: Implicit_Rating_Calculation_category.ipynb to train and test the model.

#### Installing the Spotlight library 

!conda install -c maciejkula -c pytorch spotlight --y

!conda install -c maciejkula -c pytorch -c peterjc123 spotlight=0.1.4

!pip install git+https://github.com/maciejkula/spotlight.git@master --upgrade 

# Loading needed libraries
import numpy as np
import pandas as pd
import datetime as dt
from datetime import date
import torch
from sklearn.preprocessing import LabelEncoder
import gc

# Spotlight Libraries
from spotlight.factorization.explicit import ExplicitFactorizationModel
from spotlight.interactions import Interactions
from spotlight.cross_validation import random_train_test_split
from spotlight.evaluation import rmse_score


# Loading libraries for S3 bucket connection
import boto3
import io
from io import StringIO,BytesIO, TextIOWrapper
import gzip

client = boto3.client('s3') 
resource = boto3.resource('s3') 

The expected data input should contain the user ids, the product ids, implicit rating and the timestamps (which are optional)

#### Data Preparation

# Reading Training,validation and testing dfs - Using the created T_implicit_cat data with timestamp
train_df = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/input/T_implicit_cat_rating_train.csv')
val_train_df = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/input/T_implicit_cat_rating_val_train.csv')
test_df = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/input/T_implicit_cat_rating_test.csv')
val_test_df = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/input/T_implicit_cat_rating_val_test.csv')

# Transforming event_time columns into timestamp  
train_df['event_date'] = train_df['event_time'].str[:19]# Grabbing only timestamp portion from original event_time column
train_df['event_date'] = train_df['event_date'].astype('datetime64[ns]')
train_df['timestamp'] = train_df['event_date'].values.astype(np.int64)//10**9
train_df['timestamp'] = train_df['timestamp'].astype(np.int32)
train_df.head()

# Transform category and user ids to needed format

# instantiating the labelencoder object
le = LabelEncoder()

train_df['catID'] = train_df['catID'].astype(np.int32)
train_df['userID'] = le.fit_transform(train_df['user_id'])
train_df['userID'] = train_df['userID'].astype(np.int32)

train_df.head()

train_df.nunique()

train_df.dtypes

# Applying the same to all other dfs
val_train_df['event_date'] = val_train_df['event_time'].str[:19]# Grabbing only timestamp portion from original event_time column
val_train_df['event_date'] = val_train_df['event_date'].astype('datetime64[ns]')
val_train_df['timestamp'] = val_train_df['event_date'].values.astype(np.int64)//10**9
val_train_df['timestamp'] = val_train_df['timestamp'].astype('str')

val_test_df['event_date'] = val_test_df['event_time'].str[:19]# Grabbing only timestamp portion from original event_time column
val_test_df['event_date'] = val_test_df['event_date'].astype('datetime64[ns]')
val_test_df['timestamp'] = val_test_df['event_date'].values.astype(np.int64)//10**9
val_test_df['timestamp'] = val_test_df['timestamp'].astype('str')

test_df['event_date'] = test_df['event_time'].str[:19]# Grabbing only timestamp portion from original event_time column
test_df['event_date'] = test_df['event_date'].astype('datetime64[ns]')
test_df['timestamp'] = test_df['event_date'].values.astype(np.int64)//10**9
test_df['timestamp'] = test_df['timestamp'].astype('str')

# Transform product and user ids to needed format
val_train_df['catID'] = val_train_df['catID'].astype(np.int32)
val_train_df['userID'] = le.fit_transform(val_train_df['user_id'])
val_train_df['userID'] = val_train_df['userID'].astype(np.int32)

val_train_df.nunique()

# Transform product and user ids to needed format
val_test_df['catID'] = val_test_df['catID'].astype(np.int32)
val_test_df['userID'] = le.fit_transform(val_test_df['user_id'])
val_test_df['userID'] = val_test_df['userID'].astype(np.int32)

# Transform product and user ids to needed format
test_df['catID'] = test_df['catID'].astype(np.int32)
test_df['userID'] = le.fit_transform(test_df['user_id'])
test_df['userID'] = test_df['userID'].astype(np.int32)

# Creating interaction Spotlight objects since Spotlight model expects this specific type of object
train=Interactions(user_ids=train_df['userID'].to_numpy(),item_ids=train_df['catID'].to_numpy(),ratings=train_df['implicit_rating'].to_numpy(),timestamps=train_df['timestamp'].to_numpy())

test=Interactions(user_ids=test_df['userID'].to_numpy(),item_ids=test_df['catID'].to_numpy(),ratings=test_df['implicit_rating'].to_numpy(),timestamps=test_df['timestamp'].to_numpy())

val_train=Interactions(user_ids=val_train_df['userID'].to_numpy(),item_ids=val_train_df['catID'].to_numpy(),ratings=val_train_df['implicit_rating'].to_numpy(),timestamps=val_train_df['timestamp'].to_numpy())

val_test=Interactions(user_ids=val_test_df['userID'].to_numpy(),item_ids=val_test_df['catID'].to_numpy(),ratings=val_test_df['implicit_rating'].to_numpy(),timestamps=val_test_df['timestamp'].to_numpy())

train

model = ExplicitFactorizationModel(loss='regression',
                                   embedding_dim=128,  # latent dimensionality
                                   n_iter=1,  # number of epochs of training
                                   batch_size=1024,  # minibatch size
                                   l2=1e-06,  # strength of L2 regularization
                                   learning_rate=0.1,
                                   use_cuda=1)

model.fit(train, verbose=True)# Training model

torch.save(model, 'MF_Spotlight.pt')# saving model

model = torch.load('MF_Spotlight.pt')# loading model

test_csr = test.tocsr()# converting test df to sparse row format

# Making predictions with model test
results=[]
for userID, row in enumerate(test_csr):
    predictions = model.predict(userID)
    results.append(predictions.argsort()[-10:][::-1])

# Converting prediction of model into a df
predictions_df = pd.DataFrame(data=results)
predictions_df.head()

# Storing Df with userID
results=[]
for userID, row in enumerate(test_csr):
    results.append(userID)

user_df = pd.DataFrame(data=results)
user_df.columns = ['userID']

predictions_df['userID']= user_df['userID']# mapping userID to predictions_df
predictions_df.head()

predictions_df = predictions_df.groupby('userID').tail(1)# Handling cases with multiple sequences to only grab latest one
# Rearranging recs from rows to columns
predictions_df = predictions_df.melt(id_vars=["userID"], var_name="Category_Rank", value_name="catID")
predictions_df.head()

cat_mapping = test_df[['catID','category','category_id']]
cat_mapping = cat_mapping.drop_duplicates(subset=['catID','category','category_id'])

user_mapping = test_df[['userID','user_id']]
user_mapping = user_mapping.drop_duplicates(subset=['userID','user_id'])

# Merging predictions_df to obtain the correct user and category
predictions_df = pd.merge(predictions_df, cat_mapping,  how='inner', on='catID')
predictions_df = pd.merge(predictions_df, user_mapping,  how='inner', on='userID')
#Dropping duplicates
predictions_df = predictions_df.drop_duplicates(['user_id','catID','category','category_id'])
predictions_df.head(10)

# Saving Results in S3
predictions_df.to_csv('s3://myaws-capstone-bucket/data/modeling/output/MF_Spotlight_Param2.csv',index=False)
predictions_df.nunique()
