==> Modeling/19_RankFM_model_category.ipynb <==
This notebook uses the package rankfm found at: https://github.com/etlundquist/rankfm

Also using as reference the following article: https://towardsdatascience.com/factorization-machines-for-item-recommendation-with-implicit-feedback-data-5655a7c749db

This code uses the training, validation and test sets created on the notebook: Implicit_Rating_Calculation_category.ipynb to train and test the model.

! pip install rankfm

# Loading needed libraries
import numpy as np
import pandas as pd
import datetime as dt
from datetime import date
import gc

from rankfm.rankfm import RankFM
from rankfm.evaluation import hit_rate, reciprocal_rank, discounted_cumulative_gain, precision, recall

# Loading libraries for S3 bucket connection
import boto3
import io
from io import StringIO,BytesIO, TextIOWrapper
import gzip

client = boto3.client('s3') 
resource = boto3.resource('s3') 

#### Data preparation

# Reading Training,validation and testing dfs - Using the created implicit_cat data with timestamp
train_df = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/input/T_implicit_cat_rating_train.csv')
test_df = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/input/T_implicit_cat_rating_test.csv')

# Transform product and user ids to needed format
train_df['catID'] = train_df['catID'].astype("int64")
train_df['user_id'] = train_df['user_id'].astype("int64")

# Transform product and user ids to needed format
test_df['catID'] = test_df['catID'].astype("int64")
test_df['user_id'] = test_df['user_id'].astype("int64")

#### Model Data Analysis and Training 

# RankFM two columns: a user_id and a category_id 
interactions_train = train_df[['user_id', 'catID']]
interactions_test = test_df[['user_id', 'catID']]

train_users = np.sort(interactions_train.user_id.unique())
test_users = np.sort(interactions_test.user_id.unique())
cold_start_users = set(test_users) - set(train_users)

train_items = np.sort(interactions_train.catID.unique())
test_items = np.sort(interactions_test.catID.unique())
cold_start_items = set(test_items) - set(train_items)

model = RankFM(factors=20, loss='warp', max_samples=20, alpha=0.01, sigma=0.1, learning_rate=0.1, learning_schedule='invscaling')
model.fit(interactions_train, epochs=20, verbose=True)

# Generating top 10 recommended products for each user
test_recs = model.recommend(test_users, n_items=10, filter_previous=True, cold_start='drop')

test_recs = test_recs.reset_index()# Reseting the index as the index is the user_id
test_recs.head()

# Renaming column
test_recs.columns = ['user_id', 0, 1, 2, 3, 4, 5, 6, 7, 8, 9]

# Saving results of recommendations in S3
test_recs.to_csv('s3://myaws-capstone-bucket/data/modeling/output/rankFM_category.csv',index=False)

# Transforming recommendations from being in columns to being in rows
recs = test_recs.melt(id_vars=["user_id"], 
        var_name="predicted_Rank", 
        value_name="predicted_recs")
recs['predicted_recs'] = recs['predicted_recs'].astype("int64")
recs.head()

# Saving results of recommendations in S3
recs.to_csv('s3://myaws-capstone-bucket/data/modeling/output/rankFM_T_category.csv',index=False)

recs.nunique()
