==> Modeling/21_Popularity_Based_Recommender.ipynb <==
In this code a simple popularity based recommendation model is built in order to provide recommendations based on most purchased categories to different customers. This is meant to have options for customers that we do not have a lot of information about or in cases where other recommendation systems fail to provide recommendations to customers.

Popularity Based recommendations will be provided to each customer based on their Segment which was predicted on the notebook:Customer_Segmentation_Clustering.ipynb

The idea takes inspiration from the research paper:
- Rodrigues, F., & Ferreira, B. (2016). Product recommendation based on shared Customer's Behaviour. Procedia Computer Science, 100, 136-146. doi:10.1016/j.procs.2016.09.133

# Loading basic needed libraries
import pandas as pd
import numpy as np
import gc

# Loading libraries for S3 bucket connection
import boto3
import io
from io import StringIO,BytesIO, TextIOWrapper
import gzip

client = boto3.client('s3') 
resource = boto3.resource('s3') 

# Reading clustered dataset to segment the customers based on their K-means cluster segments and perform Basket Market Analysis per Segment
# dataset came from notebook - Customer_Segmentation_Clustering.ipynb
cluster_df = pd.read_csv('s3://myaws-capstone-bucket/data/cluster_segments.csv')
cluster_df.nunique()

# Reading purchased dataframe 
purchase_df = pd.read_csv('s3://myaws-capstone-bucket/eCommerce_purchase_data.csv')
# Combining each user in the training df with their respective cluster
purchase_df = pd.merge(purchase_df,cluster_df, on=["user_id"], how='inner')
purchase_df.nunique()

purchase_df.head()

# Creating a dfaining df for each customer cluster 
df1 = purchase_df.loc[purchase_df['cluster'] == 0]
df2 = purchase_df.loc[purchase_df['cluster'] == 1]
df3 = purchase_df.loc[purchase_df['cluster'] == 2]
df4 = purchase_df.loc[purchase_df['cluster'] == 3]
df5 = purchase_df.loc[purchase_df['cluster'] == 4]
df6 = purchase_df.loc[purchase_df['cluster'] == 5]

# Counting the number of times each category was purchased for this cluster group
cluster0_recs = df1.groupby(['category', 'cluster']).size().reset_index(name = 'total_purchases')
cluster0_recs.head(10)

# Grabbing Top 10 most purchased categories for this cluster group
cluster0_recs = cluster0_recs.sort_values(['total_purchases'],ascending = False).head(10)
cluster0_recs

# Creating purchase_rank column based on total purchases
cluster0_recs['purchase_rank'] = cluster0_recs.sort_values(['total_purchases'],ascending=False).groupby('cluster').cumcount()+1
cluster0_recs['purchase_rank'] = cluster0_recs['purchase_rank'].astype('int64')
cluster0_recs = cluster0_recs[['category', 'cluster', 'purchase_rank']]
cluster0_recs

# Counting the number of times each category was purchased for this cluster group
cluster1_recs = df2.groupby(['category', 'cluster']).size().reset_index(name = 'total_purchases')

cluster1_recs = cluster1_recs.sort_values(['total_purchases'],ascending = False).head(10)

cluster1_recs['purchase_rank'] = cluster1_recs.sort_values(['total_purchases'],ascending=False).groupby('cluster').cumcount()+1
cluster1_recs['purchase_rank'] = cluster1_recs['purchase_rank'].astype('int64')
cluster1_recs = cluster1_recs[['category', 'cluster', 'purchase_rank']]
cluster1_recs

# Counting the number of times each category was purchased for this cluster group
cluster2_recs = df3.groupby(['category', 'cluster']).size().reset_index(name = 'total_purchases')

cluster2_recs = cluster2_recs.sort_values(['total_purchases'],ascending = False).head(10)

cluster2_recs['purchase_rank'] = cluster2_recs.sort_values(['total_purchases'],ascending=False).groupby('cluster').cumcount()+1
cluster2_recs['purchase_rank'] = cluster2_recs['purchase_rank'].astype('int64')
cluster2_recs = cluster2_recs[['category', 'cluster', 'purchase_rank']]
cluster2_recs

# Counting the number of times each category was purchased for this cluster group
cluster3_recs = df4.groupby(['category', 'cluster']).size().reset_index(name = 'total_purchases')

cluster3_recs = cluster3_recs.sort_values(['total_purchases'],ascending = False).head(10)

cluster3_recs['purchase_rank'] = cluster3_recs.sort_values(['total_purchases'],ascending=False).groupby('cluster').cumcount()+1
cluster3_recs['purchase_rank'] = cluster3_recs['purchase_rank'].astype('int64')
cluster3_recs = cluster3_recs[['category', 'cluster', 'purchase_rank']]
cluster3_recs

# Counting the number of times each category was purchased for this cluster group
cluster4_recs = df5.groupby(['category', 'cluster']).size().reset_index(name = 'total_purchases')

cluster4_recs = cluster4_recs.sort_values(['total_purchases'],ascending = False).head(10)

cluster4_recs['purchase_rank'] = cluster4_recs.sort_values(['total_purchases'],ascending=False).groupby('cluster').cumcount()+1
cluster4_recs['purchase_rank'] = cluster4_recs['purchase_rank'].astype('int64')
cluster4_recs = cluster4_recs[['category', 'cluster', 'purchase_rank']]
cluster4_recs

# Counting the number of times each category was purchased for this cluster group
cluster5_recs = df6.groupby(['category', 'cluster']).size().reset_index(name = 'total_purchases')

cluster5_recs = cluster5_recs.sort_values(['total_purchases'],ascending = False).head(10)

cluster5_recs['purchase_rank'] = cluster5_recs.sort_values(['total_purchases'],ascending=False).groupby('cluster').cumcount()+1
cluster5_recs['purchase_rank'] = cluster5_recs['purchase_rank'].astype('int64')
cluster5_recs = cluster5_recs[['category', 'cluster', 'purchase_rank']]
cluster5_recs

# Stacking cluster recs together with concat function
data_frames = [cluster0_recs, cluster1_recs, cluster2_recs, cluster3_recs, cluster4_recs, cluster5_recs]
top_recs_df = pd.concat(data_frames)
top_recs_df.nunique()

top_recs_df.head()

# Reading the test users and mapping each user to their cluster
test_df = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/input/implicit_cat_rating_test.csv')
test_df = pd.merge(test_df,cluster_df, on=["user_id"], how='inner')
test_df.nunique()

# Mapping each user to their top 10 popular recs based on their cluster
popular_recs_df = pd.merge(test_df[['user_id','cluster']],top_recs_df, on=["cluster"], how='inner')
popular_recs_df.nunique()

popular_recs_df = popular_recs_df.drop_duplicates(subset=['user_id', 'category', 'purchase_rank'])
popular_recs_df.head(50)

popular_recs_df.tail(50)

# Saving results of recommendations in S3
popular_recs_df.to_csv('s3://myaws-capstone-bucket/data/modeling/output/popular_recs.csv',index=False)
