==> Modeling/22_Model_Performance_Testing_MAP@K.ipynb <==
This code makes use of the MAP@K Rank Accuracy Metric to evaluate the created models 

! pip install ml_metrics

# Loading basic needed libraries
import pandas as pd
import numpy as np
import gc
import ml_metrics as metrics

# Loading libraries for S3 bucket connection
import boto3
import io
from io import StringIO,BytesIO, TextIOWrapper
import gzip

client = boto3.client('s3') 
resource = boto3.resource('s3') 

#### Reading the testing dataset

- All models will be tested for accuracy and compared with the same test dataset so that the results are comparable

test_df = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/input/implicit_cat_rating_test.csv')
test_df.head()

test_df.shape

test_df.nunique()

# Creating rank columns for implicit rating
test_df['category_rank'] = test_df.sort_values(['implicit_rating'],ascending=False).groupby('user_id').cumcount()+1
test_df = test_df.sort_values(['user_id','category_rank'],ascending=True)
test_df.head()

users_2_test = test_df[['user_id']]
users_2_test = users_2_test.drop_duplicates(subset=['user_id'])
users_2_test.nunique()

#### Testing Popularity Based Recommender

pop_recs = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/output/popular_recs.csv')
pop_recs.head()

pop_recs.nunique()

pop_recs.shape

pop_recs = pop_recs.sort_values(['user_id','purchase_rank'],ascending=True)
pop_recs.head()

Making sure output is as expected for mapk function 

Using the following article as reference: https://www.kaggle.com/wendykan/map-k-demo/notebook?select=test.csv

#Transforming categories for both recs and actual categories to lists for each user
predicted = pop_recs.groupby('user_id')['category'].apply(list)
actual = test_df.groupby('user_id')['category'].apply(list)

predicted

actual

# Obtaining mapk value from popular recs vs actual test dataset comparrison 
mapk_score = metrics.mapk(actual, predicted, 10)
print('MAP@10:','{:.8f}'.format(mapk_score))

mapk_score = metrics.mapk(actual, predicted, 5)
print('MAP@5:','{:.8f}'.format(mapk_score))

#### Testing MF Spotlight Recommender

MF_recs = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/output/MF_Spotlight_Param1.csv')
MF_recs.head()

MF_recs.nunique()

MF_recs = MF_recs.sort_values(['user_id','Category_Rank'],ascending=True)
MF_recs.head()

MF_recs.shape

test_df = test_df.sort_values(['user_id','category_rank'],ascending=True)
test_df.head()

test_df.shape

#Transforming categories for both recs and actual categories to lists for each user
predicted = MF_recs.groupby('user_id')['category'].apply(list)
actual = test_df.groupby('user_id')['category'].apply(list)

# Obtaining mapk value from MF recs vs actual test dataset comparrison
mapk_score = metrics.mapk(actual, predicted, 10) 
print('MAP@10:','{:.8f}'.format(mapk_score))

mapk_score = metrics.mapk(actual, predicted, 5) 
print('MAP@5:','{:.8f}'.format(mapk_score))

MF_recs = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/output/MF_Spotlight_Param2.csv')
MF_recs = MF_recs.sort_values(['user_id','Category_Rank'],ascending=True)
MF_recs.shape

MF_recs.head(15)

test_df.head()

#Transforming categories for both recs and actual categories to lists for each user
predicted = MF_recs.groupby('user_id')['category'].apply(list)
actual = test_df.groupby('user_id')['category'].apply(list)

# Obtaining mapk value from MF recs vs actual test dataset comparrison
mapk_score = metrics.mapk(actual, predicted, 10) 
print('MAP@10:','{:.8f}'.format(mapk_score))

mapk_score = metrics.mapk(actual, predicted, 5) 
print('MAP@5:','{:.8f}'.format(mapk_score))

MF_recs = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/output/MF_Spotlight_Param3.csv')
MF_recs = MF_recs.sort_values(['user_id','Category_Rank'],ascending=True)
MF_recs.shape

#Transforming categories for both recs and actual categories to lists for each user
predicted = MF_recs.groupby('user_id')['category'].apply(list)
actual = test_df.groupby('user_id')['category'].apply(list)

# Obtaining mapk value from MF recs vs actual test dataset comparrison
mapk_score = metrics.mapk(actual, predicted, 10) 
print('MAP@10:','{:.8f}'.format(mapk_score))

mapk_score = metrics.mapk(actual, predicted, 5) 
print('MAP@5:','{:.8f}'.format(mapk_score))

#### Testing LSTM - RNN Model

LSTM_recs = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/output/LSTM_param1.csv')
LSTM_recs.head()

LSTM_recs.nunique()

LSTM_recs = LSTM_recs.sort_values(['user_id','Category_Rank'],ascending=True)
LSTM_recs.head()

LSTM_recs.shape

# Grabbing ids to use for test_df to verify performance of LSTM model
lstm_users = LSTM_recs[['user_id']]
lstm_users.drop_duplicates(subset=['user_id'])
lstm_users.nunique()

actual = pd.merge(test_df,lstm_users, on=["user_id"], how='inner')
actual = actual.drop_duplicates(subset=['user_id', 'category', 'category_id', 'implicit_rating', 'catID'])
actual = actual.sort_values(['user_id','category_rank'],ascending=True)
actual.head()

actual.nunique()

actual.shape

#Transforming categories for both recs and actual categories to lists for each user
predicted = LSTM_recs.groupby('user_id')['category'].apply(list)
actual = actual.groupby('user_id')['category'].apply(list)

# Obtaining mapk value from LSTM recs vs actual test dataset comparrison 
mapk_score = metrics.mapk(actual, predicted, 10)
print('MAP@10:','{:.8f}'.format(mapk_score))

mapk_score = metrics.mapk(actual, predicted, 5)
print('MAP@5:','{:.8f}'.format(mapk_score))

LSTM_recs = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/output/LSTM_param2.csv')
LSTM_recs = LSTM_recs.sort_values(['user_id','Category_Rank'],ascending=True)
LSTM_recs.shape

# Grabbing ids to use for test_df to verify performance of LSTM model
lstm_users = LSTM_recs[['user_id']]
lstm_users.drop_duplicates(subset=['user_id'])

actual = pd.merge(test_df,lstm_users, on=["user_id"], how='inner')
actual = actual.drop_duplicates(subset=['user_id', 'category', 'category_id', 'implicit_rating', 'catID'])
actual = actual.sort_values(['user_id','category_rank'],ascending=True)
actual.shape

#Transforming categories for both recs and actual categories to lists for each user
predicted = LSTM_recs.groupby('user_id')['category'].apply(list)
actual = actual.groupby('user_id')['category'].apply(list)

# Obtaining mapk value from LSTM recs vs actual test dataset comparrison 
mapk_score = metrics.mapk(actual, predicted, 10)
print('MAP@10:','{:.8f}'.format(mapk_score))

mapk_score = metrics.mapk(actual, predicted, 5)
print('MAP@5:','{:.8f}'.format(mapk_score))

LSTM_recs = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/output/LSTM_param3.csv')
LSTM_recs = LSTM_recs.sort_values(['user_id','Category_Rank'],ascending=True)
LSTM_recs.shape

# Grabbing ids to use for test_df to verify performance of LSTM model
lstm_users = LSTM_recs[['user_id']]
lstm_users.drop_duplicates(subset=['user_id'])

actual = pd.merge(test_df,lstm_users, on=["user_id"], how='inner')
actual = actual.drop_duplicates(subset=['user_id', 'category', 'category_id', 'implicit_rating', 'catID'])
actual = actual.sort_values(['user_id','category_rank'],ascending=True)
actual.shape

#Transforming categories for both recs and actual categories to lists for each user
predicted = LSTM_recs.groupby('user_id')['category'].apply(list)
actual = actual.groupby('user_id')['category'].apply(list)

# Obtaining mapk value from LSTM recs vs actual test dataset comparrison 
mapk_score = metrics.mapk(actual, predicted, 10)
print('MAP@10:','{:.8f}'.format(mapk_score))

mapk_score = metrics.mapk(actual, predicted, 5)
print('MAP@5:','{:.8f}'.format(mapk_score))

#### Testing CNN Model

CNN_recs = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/output/CNN_param1.csv')
CNN_recs.head()

CNN_recs.nunique()

CNN_recs = CNN_recs.sort_values(['user_id','Category_Rank'],ascending=True)
CNN_recs.head()

CNN_recs.shape

# Grabbing ids to use for test_df to verify performance of cnn model
cnn_users = CNN_recs[['user_id']]
cnn_users.drop_duplicates(subset=['user_id'])
cnn_users.nunique()

actual = pd.merge(test_df,cnn_users, on=["user_id"], how='inner')
actual = actual.drop_duplicates(subset=['user_id', 'category', 'category_id', 'implicit_rating', 'catID'])
actual = actual.sort_values(['user_id','category_rank'],ascending=True)
actual.head()

actual.nunique()

actual.shape

#Transforming categories for both recs and actual categories to lists for each user
predicted = CNN_recs.groupby('user_id')['category'].apply(list)
actual = actual.groupby('user_id')['category'].apply(list)

# Obtaining mapk value from CNN recs vs actual test dataset comparrison 
mapk_score = metrics.mapk(actual, predicted, 10) 
print('MAP@10:','{:.8f}'.format(mapk_score))

mapk_score = metrics.mapk(actual, predicted, 5) 
print('MAP@5:','{:.8f}'.format(mapk_score))

CNN_recs = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/output/CNN_param2.csv')
CNN_recs = CNN_recs.sort_values(['user_id','Category_Rank'],ascending=True)
CNN_recs.shape

# Grabbing ids to use for test_df to verify performance of cnn model
cnn_users = CNN_recs[['user_id']]
cnn_users.drop_duplicates(subset=['user_id'])

actual = pd.merge(test_df,cnn_users, on=["user_id"], how='inner')
actual = actual.drop_duplicates(subset=['user_id', 'category', 'category_id', 'implicit_rating', 'catID'])
actual = actual.sort_values(['user_id','category_rank'],ascending=True)
actual.shape

#Transforming categories for both recs and actual categories to lists for each user
predicted = CNN_recs.groupby('user_id')['category'].apply(list)
actual = actual.groupby('user_id')['category'].apply(list)

# Obtaining mapk value from CNN recs vs actual test dataset comparrison 
mapk_score = metrics.mapk(actual, predicted, 10) 
print('MAP@10:','{:.8f}'.format(mapk_score))

mapk_score = metrics.mapk(actual, predicted, 5) 
print('MAP@5:','{:.8f}'.format(mapk_score))

CNN_recs = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/output/CNN_param3.csv')
CNN_recs = CNN_recs.sort_values(['user_id','Category_Rank'],ascending=True)
CNN_recs.shape

# Grabbing ids to use for test_df to verify performance of cnn model
cnn_users = CNN_recs[['user_id']]
cnn_users.drop_duplicates(subset=['user_id'])

actual = pd.merge(test_df,cnn_users, on=["user_id"], how='inner')
actual = actual.drop_duplicates(subset=['user_id', 'category', 'category_id', 'implicit_rating', 'catID'])
actual = actual.sort_values(['user_id','category_rank'],ascending=True)
actual.shape

#Transforming categories for both recs and actual categories to lists for each user
predicted = CNN_recs.groupby('user_id')['category'].apply(list)
actual = actual.groupby('user_id')['category'].apply(list)

# Obtaining mapk value from CNN recs vs actual test dataset comparrison 
mapk_score = metrics.mapk(actual, predicted, 10) 
print('MAP@10:','{:.8f}'.format(mapk_score))

mapk_score = metrics.mapk(actual, predicted, 5) 
print('MAP@5:','{:.8f}'.format(mapk_score))

#### Testing RankFM model

RankFM_recs = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/output/rankFM_T_category.csv')
RankFM_recs.columns = ['user_id','predicted_Rank','catID']
RankFM_recs.head()

RankFM_recs = RankFM_recs.sort_values(['user_id','predicted_Rank'],ascending=True)
RankFM_recs.head()

RankFM_recs.nunique()

RankFM_recs.shape

# Grabbing ids to use for test_df to verify performance of RankFM model
RankFM_users = RankFM_recs[['user_id']]
RankFM_users.drop_duplicates(subset=['user_id'])
RankFM_users.nunique()

actual = pd.merge(test_df,RankFM_users, on=["user_id"], how='inner')
actual = actual.drop_duplicates(subset=['user_id', 'category', 'category_id', 'implicit_rating', 'catID'])
actual = actual.sort_values(['user_id','category_rank'],ascending=True)
actual.head()

actual.nunique()

actual.shape

RankFM_recs.head()

predicted = pd.merge(actual[['user_id']],RankFM_recs, on=["user_id"], how='inner')

predicted = predicted.drop_duplicates(subset=['user_id', 'predicted_Rank', 'catID'])
predicted = predicted.sort_values(['user_id','predicted_Rank'],ascending=True)
predicted.head()

predicted.nunique()

predicted.shape

#Transforming categories for both recs and actual categories to lists for each user
predicted = predicted.groupby('user_id')['catID'].apply(list)
actual = actual.groupby('user_id')['catID'].apply(list)

# Obtaining mapk value from RankFM recs vs actual test dataset comparrison 
mapk_score = metrics.mapk(actual, predicted, 10) 
print('MAP@10:','{:.8f}'.format(mapk_score))

mapk_score = metrics.mapk(actual, predicted, 5) 
print('MAP@5:','{:.8f}'.format(mapk_score))

#### Testing CF Model

# Reading CF results
CF_recs = pd.read_csv('s3://myaws-capstone-bucket/data/modeling/ALS/results/part-00000-4e8dadd0-ae3c-4650-a762-ccf1cf39864a-c000.csv')
CF_recs.columns = ['user_id','catID','predicted_rank','predicted_rating']
# Keeping only test users for model evaluation
CF_recs = pd.merge(users_2_test,CF_recs, on=["user_id"], how='inner')
CF_recs = CF_recs.drop_duplicates(subset=['user_id','catID','predicted_rank','predicted_rating'])
CF_recs = CF_recs.sort_values(['user_id','predicted_rank'],ascending=True)
CF_recs.head()

CF_recs.nunique()

# Keeping same amount of users on test data
actual = pd.merge(test_df,CF_recs[['user_id']], on=["user_id"], how='inner')
actual = actual.drop_duplicates(subset=['user_id', 'category', 'category_id', 'implicit_rating', 'catID'])
actual = actual.sort_values(['user_id','category_rank'],ascending=True)
actual.head()

actual.nunique()

#Transforming categories for both recs and actual categories to lists for each user
predicted = CF_recs.groupby('user_id')['catID'].apply(list)
actual = actual.groupby('user_id')['catID'].apply(list)

# Obtaining mapk value from RankFM recs vs actual test dataset comparrison 
mapk_score = metrics.mapk(actual, predicted, 10) 
print('MAP@10:','{:.8f}'.format(mapk_score))

mapk_score = metrics.mapk(actual, predicted, 5) 
print('MAP@5:','{:.8f}'.format(mapk_score))
